Implemented Oath Github, added login functionality, improved accesibility

+++ server.js
@@ -3,9 +3,83 @@ require("dotenv").config()
 const express = require("express");
 const path = require("path");
 const { MongoClient } = require("mongodb");
+const session = require("express-session");
+const passport = require("passport");
+const GitHubStrategy = require("passport-github2").Strategy;
 
 const app = express();
 const PORT = process.env.PORT || 3000;
+
+// information we put into the session to represent this user
+passport.serializeUser(function(user, done) {
+    done(null, user); // put the user object into the session
+});
+
+// when a request comes in with a session, turn the stored session information back into a user.
+passport.deserializeUser(function(user, done) {
+    done(null, user);
+});
+
+// Oauth w github authentication
+passport.use(new GitHubStrategy({
+    clientID: process.env.GITHUB_CLIENT_ID,
+    clientSecret: process.env.GITHUB_CLIENT_SECRET,
+    callbackURL: "/auth/github/callback"
+}, function(accessToken, refreshToken, profile, done) {
+    done(null, profile);
+}));
+
+// use sessions for this application
+app.use(session({
+    secret: process.env.SESSION_SECRET,
+    resave: false,
+    saveUninitialized: false
+}));
+
+// use passport for authentication
+app.use(passport.initialize());
+app.use(passport.session());
+
+
+const requireLogin = function(request, response, next) {
+    if (request.isAuthenticated()) {
+        next();
+    }
+    else {
+        response.redirect("/");
+    }
+};
+
+//when someone visits it, passport sends them to github to log in
+app.get("/auth/github",
+    passport.authenticate("github", { scope: ["user:email"] })
+);
+
+app.get("/auth/github/callback",
+    passport.authenticate("github", {
+        failureRedirect: "/"
+    }),
+    function(request, response) {
+        response.redirect("/");
+    }
+);
+
+app.get("/logout", function(request, response) {
+    request.logout(function(error) {
+        if (error) {
+            return response.status(500).send("Logout failed");
+        }
+
+        response.redirect("/");
+    });
+});
+
+app.get("/auth/status", function(request, response) {
+    response.json({
+        loggedIn: request.isAuthenticated()
+    });
+});
+
 
 const client = new MongoClient(process.env.MONGODB_URI);
 const db = client.db("a3-grocery");
@@ -23,7 +97,17 @@ client.connect()
 app.use(express.json());
 
 // Serve files from the public folder (GET)
-app.use(express.static(path.join(__dirname, "public")));
+// false stops Express from automatically serving index.html at /
+app.use(express.static(path.join(__dirname, "public"), { index: false }));
+
+app.get("/", function(request, response) {
+    if (request.isAuthenticated()) {
+        response.sendFile(path.join(__dirname, "public", "index.html"));
+    }
+    else {
+        response.sendFile(path.join(__dirname, "public", "login.html"));
+    }
+});
 
 // Start server
 app.listen(PORT, () => {
@@ -31,24 +115,24 @@ app.listen(PORT, () => {
 });
 
 //ENDPOINTS
-app.post("/submit", function(request, response) {
+app.post("/submit", requireLogin, function(request, response) {
   console.log(request.headers);
   console.log(request.body);
   handleSubmit(request, response);
 });
 
-app.post("/update", function(request, response) {
+app.post("/update", requireLogin, function(request, response) {
     console.log(request.headers);
     console.log(request.body);
     handleUpdate(request, response);
 })
 
-app.post("/delete", function(request, response) {
+app.post("/delete", requireLogin, function(request, response) {
   handleDelete(request, response);
 })
 
-app.get("/items", async function(request, response) {
-    const groceryList = await items.find().toArray()
+app.get("/items",requireLogin,  async function(request, response) {
+    const groceryList = await items.find({ userId: request.user.id }).toArray()
     response.status(200).json(groceryList)
 })
 
@@ -62,6 +146,7 @@ const handleSubmit = async function( request, response ) {
   console.log(request.body);
 
     const listItem = request.body
+    listItem.userId = request.user.id // attaches user id to item before MongoDB saves it
 
     // add a unique id to the item
     const lastItem = await items.find().sort({ id: -1 }).limit(1).toArray();
@@ -72,7 +157,7 @@ const handleSubmit = async function( request, response ) {
 
     await items.insertOne( listItem )
 
-    const groceryList = await items.find().toArray()
+    const groceryList = await items.find({ userId: request.user.id }).toArray()
 
     response.status(200).json(groceryList)
 }
@@ -81,11 +166,11 @@ const handleUpdate = async function (request, response) {
     const update = request.body
 
     await items.updateOne(
-        { id: update.id }, // find the item that matches the id of the checked/unchecked item
+        { id: update.id, userId: request.user.id }, // find the item that matches the id of the checked/unchecked item
         { $set: { is_purchased: update.is_purchased } } // update the is_purchased property of the item
     )
 
-    const groceryList = await items.find().toArray()
+    const groceryList = await items.find({ userId: request.user.id }).toArray()
 
     console.log(groceryList)
 
@@ -96,9 +181,9 @@ const handleDelete = async function (request, response) {
 
     const deleted = request.body
 
-    await items.deleteOne({ id: deleted.id })
+    await items.deleteOne({ id: deleted.id, userId: request.user.id })
 
-    const groceryList = await items.find().toArray()
+    const groceryList = await items.find({ userId: request.user.id }).toArray()
 
     console.log(groceryList)
